+++ src/bpms/bpmn/elements/BusinessRuleTask.ts
@@ -1,0 +1,5 @@
+// tslint:disable-next-line: no-var-requires
+const {SignalTask : BusinessRuleTask}  = require("bpmn-elements");
+
+export default BusinessRuleTask;
+