+++ src/bpms/cmmn/CmmnEngine.ts
@@ -1,4 +1,5 @@
 import { uuidv1 } from "nowjs-core/lib/utils/UuidUtils";
+import { BpmsEngine } from "../BpmsEngine";
 
 export interface CmmnEngineOptions {
   name: string;
@@ -9,10 +10,16 @@ export class CmmnEngine {
   private id: string = uuidv1();
   private name: string;
   private options: CmmnEngineOptions;
-  public static createEngine(options?: CmmnEngineOptions): CmmnEngine {
-    return new CmmnEngine(options);
+
+  public static createEngine(options?: CmmnEngineOptions): CmmnEngine;
+  public static createEngine(bpmsEngine: BpmsEngine, options?: CmmnEngineOptions): CmmnEngine;
+  public static createEngine(arg1?: BpmsEngine | CmmnEngineOptions, arg2?: CmmnEngineOptions): CmmnEngine {
+    if (arg1 instanceof BpmsEngine) {
+      return new CmmnEngine(arg1, arg2);
+    }
+    return new CmmnEngine(undefined, arg1);
   }
-  constructor(options?: CmmnEngineOptions) {
+  constructor(private bpmsEngine?: BpmsEngine, options?: CmmnEngineOptions) {
     this.options = options || { name: "CmmnEngine-" + this.id };
     this.name = this.options.name;
   }
@@ -22,5 +29,8 @@ export class CmmnEngine {
 
   public get Name(): string {
     return this.name;
+  }
+  public get BpmsEngine(): BpmsEngine | undefined {
+    return this.bpmsEngine;
   }
 }