+++ test/bpms/bpmn/BpmnProcessInstance.test.ts
@@ -1,5 +1,6 @@
 import "jest";
 import { BpmnProcessInstance } from "../../../src";
+import { source2 } from "./sampleSources";
 
 // tslint:disable:no-empty
 beforeAll(() => {});
@@ -43,39 +44,3 @@ describe("BpmnProcess", () => {
   });
 });
 
-const source1 = `
-<?xml version="1.0" encoding="UTF-8"?>
-  <definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
-               xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <process id="theProcess2" isExecutable="true">
-    <startEvent id="theStart" />
-    <exclusiveGateway id="decision" default="flow2" />
-    <endEvent id="end1" />
-    <endEvent id="end2" />
-    <sequenceFlow id="flow1" sourceRef="theStart" targetRef="decision" />
-    <sequenceFlow id="flow2" sourceRef="decision" targetRef="end1" />
-    <sequenceFlow id="flow3" sourceRef="decision" targetRef="end2">
-      <conditionExpression>true</conditionExpression>
-    </sequenceFlow>
-  </process>
-</definitions>`;
-
-const source2 = `
-<?xml version="1.0" encoding="UTF-8"?>
-<definitions xmlns="http://www.omg.org/spec/BPMN/20100524/MODEL"
-             xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
-  <process id="theProcess" isExecutable="true">
-    <dataObjectReference id="inputFromUserRef" dataObjectRef="inputFromUser" />
-    <dataObject id="inputFromUser" />
-    <startEvent id="theStart" />
-    <userTask id="userTask">
-      <ioSpecification id="inputSpec">
-        <dataOutput id="userInput" name="sirname" />
-      </ioSpecification>
-      <dataOutputAssociation id="associatedWith" sourceRef="userInput" targetRef="inputFromUserRef" />
-    </userTask>
-    <endEvent id="theEnd" />
-    <sequenceFlow id="flow1" sourceRef="theStart" targetRef="userTask" />
-    <sequenceFlow id="flow2" sourceRef="userTask" targetRef="theEnd" />
-  </process>
-</definitions>`;